requirements.txt
customtkinter
Pillow
flask
gunicorn
SQLAlchemy
psycopg2-binary
